now debugging new compactions

+++ algorithm/lsmftl/design_knob/lsmtree_param_module.h
@@ -7,7 +7,7 @@ enum LSM_TYPE{
 	TIER, LEVEL, HYBRIDTL, HYBRIDLT
 };
 
-uint32_t get_size_factor(uint32_t level, uint32_t mapping_num);
+double get_size_factor(uint32_t level, uint32_t mapping_num);
 uint32_t get_level(uint32_t sizefactor, uint32_t mapping_num);
 uint32_t get_waf(char t, uint32_t level, uint32_t size_factor);
 uint32_t get_raf(char t, uint32_t level, uint32_t size_factor);
